requirements.txt
asgiref==3.2.10
Django==3.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.42
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
